Move applyImportedComponentsToGrid
